+++ functionnal_tests/cypress/integration/notification_wall/mark_as_read_spec.js
@@ -1,4 +1,4 @@
-import { PAGES } from '../../support/urls_commands'
+import { PAGES } from '../../support/urls_commands.js'
 
 const markAllAsReadButton = '[data-cy=markAllAsReadButton]'